pyproject.toml
[tool.poetry]
name = "google-workspace-posture"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.127.0"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
argparse = "^1.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
google-workspace-posture = "google_workspace_posture.main:main"
